requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pillow
numpy
scikit-learn
pandas
tqdm
pyyaml
opencv-python
